net/http: create exported Trace function
This change factors out handler tracing into a new Trace function,
simplifying other contrib implementations such as gorilla/mux.
